+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+### 6.2.2
+
 * Fix `pause_active?` for stripe incorrectly returning `true`
 * Refactor Braintree cancel / cancel_now to use sync